Porównywanie katalogów. Sprawdzanie rozmiaru plików.

+++ filesync.h
@@ -6,6 +6,6 @@
 
 bool is_directory(const char *path);
 bool path_contains(const char *path1, const char *path2);
-void run_filesync(const char* src, const char* dst, bool is_recursive, ssize_t size_threshold);
+void run_filesync(const char* src, const char* dst, bool is_recursive, off_t size_threshold);
 
 #endif
